split image webhook gateway by responsibility
